Drop the Code AST in favour of an annotated Exp tree
Pattern synonyms are used to minimize the impedance.

The benefit is less duplication across the Exp and Code trees, which will
become even more important when implementing the solution for issue #1.

The downside is that the Exp Code is too "big" in that there exist
values our code will never generate, so we end up with a bunch of partial
functions which ignore those.